# api/api.yaml
openapi: 3.1.0
info:
  title: Polygon ID - Issuer
  description: |
    Documentation for the Issuer
  version: "1"
servers:
  - description: Local
    url: http://localhost:3001
...
paths:
...
  /v1/identities:
    post:
      summary: Create Identity
      operationId: CreateIdentity
      description: Endpoint to create an identity
      tags:
        - Identity
      security:
        - basicAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIdentityRequest'
      responses:
        '201':
          description: Identity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIdentityResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500-CreateIdentity'
    get:
      summary: Get Identities
      operationId: GetIdentities
      description: Endpoint to get all the identities
      tags:
        - Identity
      security:
        - basicAuth: [ ]
      responses:
        '200':
          description: all good
          content:
            application/json:
              schema:
                type: array
                x-omitempty: false
                items:
                  type: string
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
  /v1/{identifier}/state/retry:
    post:
      summary: Retry Publish Identity State
      operationId: RetryPublishState
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
      tags:
        - Identity
      responses:
        '202':
          description: Publish state successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishIdentityStateResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /v1/{identifier}/state/publish:
    post:
      summary: Publish Identity State
      operationId: PublishIdentityState
      description: Endpoint to publish identity state
      tags:
        - Identity
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
      responses:
        '200':
          description: No transactions to process to the given identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorMessage'
        '202':
          description: Transaction ID of the published  state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishIdentityStateResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'

  #claims:
  /v1/{identifier}/claims:
    post:
      summary: Create Claim
      operationId: CreateClaim
      description: Endpoint to create a Claim
      tags:
        - Claim
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClaimRequest'
      responses:
        '201':
          description: Claim created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateClaimResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get Claims
      operationId: GetClaims
      description: |
        Endpoint to retrieve claims 
        > ⚠️ **self** and **subject** filter cannot be used together
      tags:
        - Claim
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
        - in: query
          name: schemaType
          schema:
            type: string
          description: Filter per schema type. Example - KYCAgeCredential
        - in: query
          name: schemaHash
          schema:
            type: string
          description: Filter per schema hash. Example - c9b2370371b7fa8b3dab2a5ba81b6838
        - in: query
          name: subject
          schema:
            type: string
          description: Filter per subject. Example - did:polygonid:polygon:mumbai:2qE1BZ7gcmEoP2KppvFPCZqyzyb5tK9T6Gec5HFANQ
        - in: query
          name: revoked
          schema:
            type: boolean
          description: Filter per claims revoked or not - Example - true.
        - in: query
          name: self
          schema:
            type: boolean
          description: Filter per retrieve claims of the provided identifier. Example - true
        - in: query
          name: query_field
          schema:
            type: string
          description: Filter this field inside the data of the claim
        - in: query
          name: query_value
          schema:
            type: string
          description: Filter this value inside the data of the claim for the specified field in query_field
      responses:
        '200':
          description: Claims found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetClaimsResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
  /v1/{identifier}/claims/{id}:
    get:
      summary: Get Claim
      operationId: GetClaim
      description: Endpoint to retrieve a created Claim
      tags:
        - Claim
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
        - $ref: '#/components/parameters/pathClaim'
      responses:
        '200':
          description: Claim found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetClaimResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  /v1/{identifier}/claims/revoke/{nonce}:
    post:
      summary: Revoke Claim
      operationId: RevokeClaim
      description: Endpoint to revoke a claim
      tags:
        - Claim
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
        - $ref: '#/components/parameters/pathNonce'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeClaimResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  /v1/{identifier}/claims/revocation/status/{nonce}:
    get:
      summary: Get Revocation Status
      operationId: GetRevocationStatus
      description: Endpoint to get the revocation status
      tags:
        - Claim
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
        - $ref: '#/components/parameters/pathNonce'
        - $ref: '#/components/parameters/pathStateHash'
      responses:
        '200':
          description: Proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevocationStatusResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /v1/{identifier}/claims/{id}/qrcode:
    get:
      summary: Get Claim QR code
      operationId: GetClaimQrCode
      description: Returns a json that can be used to create the QR Code to scan for accepting a claim.
      tags:
        - Claim
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
        - $ref: '#/components/parameters/pathClaim'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetClaimQrCodeResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  /v1/claims/{id}/mtp:
    get:
      summary: Get Claim MTP
      operationId: GetClaimMTP
      description: Returns claim's MT inclusion proof for a given state hash.
      tags:
        - Claim
      parameters:
        - $ref: '#/components/parameters/pathClaim'
        - $ref: '#/components/parameters/pathStateHash'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetClaimMTPResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  /v1/{identifier}/claims/{id}/status/subscribe:
    get:
      summary: Websocket for getting claim status
      operationId: SubscribeToClaimWebsocket
      description: Subscribes to websocket for getting claim status
      tags:
        - Claim
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pathIdentifier'
        - $ref: '#/components/parameters/pathClaim'
      responses:
        '200':
          description: ok
        '500':
          $ref: '#/components/responses/500'

#agent
...
  /v1/users:
    post:
      summary: Add user
      operationId: AddUser
      description: Adds login, password and did of a user to the database.
      tags:
        - User
      security:
        - basicAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetClaimQrCodeResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    CreateIdentityRequest:
      type: object
      required:
        - didMetadata
      properties:
        didMetadata:
          type: object
          required:
            - method
            - blockchain
            - network
          properties:
            method:
              type: string
              x-omitempty: false
              example: "polygonid"
            blockchain:
              type: string
              x-omitempty: false
              example: "polygon"
            network:
              type: string
              x-omitempty: false
              example: "mumbai"
    CreateIdentityResponse:
      type: object
      required:
        - identity
        - status
      properties:
        identifier:
          type: string
        state:
          $ref: '#/components/schemas/IdentityState'
...
    PublishIdentityStateResponse:
      type: object
      properties:
        txID:
          type: string
        claimsTreeRoot:
          type: string
        state:
          type: string
        revocationTreeRoot:
          type: string
        rootOfRoots:
          type: string

    #claims
...
    CreateClaimResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          x-omitempty: false
...
  parameters:
    pathIdentifier:
      name: identifier
      in: path
      required: true
      description: Issuer identifier
      schema:
        type: string
...
  responses:
    '400':
      description: 'Bad Request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'
    '401':
      description: 'Unauthorized'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'
...
    '500':
      description: 'Internal Server error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'
    '500-CreateIdentity':
      description: 'Internal Server error'
      content:
        application/json:
          schema:
            type: object
            properties:
              requestID:
                type: string
              code:
                type: integer
              error:
                type: string